requirements.txt
aiogram==3.13.1
motor==3.0.0
python-dotenv==1.01
mistralai==1.1.0